requirements.txt
ccxt>=4.2.0
python-telegram-bot>=20.7
pandas>=2.1.0
numpy>=1.26.0
ta>=0.11.0
matplotlib>=3.8.0
mplfinance>=0.12.0
schedule>=1.2.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0
